created a file in utils for the function reusability
